+++ API/users/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class EcommerceapiConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'EcommerceApis'
+    name = 'users'
