+++ API/users/v1/serializers.py
@@ -1,7 +1,7 @@
 from dataclasses import fields
 from rest_framework import serializers
 
-from EcommerceApis.models import User
+from users.models import User
 
 
 class UserSerializer(serializers.ModelSerializer):
