+++ API/users/v1/views.py
@@ -1,11 +1,8 @@
-from EcommerceApis.models import User
+from users.models import User
 from .serializers import UserSerializer
 from django.shortcuts import render
-from rest_framework.views import APIView
 from django.contrib.auth.hashers import make_password
 from rest_framework import generics
-from rest_framework.response import Response
-from rest_framework import status
 # Create your views here.
 
 class UserRetrieveUpdateDestroy(generics.RetrieveUpdateDestroyAPIView):
